Definable kerf widths and improved small hole detection.
Was requested to make the kerf width an adjustable parameter, since it will
vary based on manufacturing process.  Now, if not set, it will fall back on the
default sizing rules from before.

And now I'm mixing two things in one commit, but it was just so easy to improve
small hole detection by adding in the new is_concave function, so I did.  This
fixes small_outside_radii.