Extra test for good internal corners.
I am about to start addressing issues with internal corners, but I realized I
never made any tests to ensure that I don't go too far with that.  So here's a
flat piece with an internal corner with a radius large enough that it should
not have any manufacturing issues.