Handle esoteric surface types.
Previously, all parts were designated "non-uniform" if they had any surface
types beyond planes, cylinders, and cones.  This now adds a heuristic that will
accept other face types as long as appropriate normal vectors are found over a
bunch of points on the surface.  Spheres and toroids are still rejected
automatically though, because it's impossible for them to  be uniform.

Also added a test showing a bad version of one of those other face types.  Had
previously added two good versions, so this extra bad one is to make sure it's
caught correctly.