Tweak to sharp internal corners to be more forgiving.
Previously called it an internal corner if the test point was within the limits
of any connected surface.  This changes it to be an internal corner only if the
point is within all connected surfaces.  The previous form was causing a
somewhat spurious extra tight-corner on nonuniform.STEP.  The new form only has
one (correct) tight-corner issue on that file.