Faster checks for small cuts.
The code to check for small cuts was very slow before due to many calls to
distToShape, which is an expensive function.  Now, checks are done on the
bounding boxes of the faces first; if the bounding boxes are farther than the
kerf width, then there is definitely no issue and no need to check precise
distance.  This has dramatically dropped the amount of time needed to run the
tests, close to where it was before the new small cuts checks were implemented.